# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "domain.BalanceResponse": {
                "properties": {
                    "balances": {
                        "items": {
                            "$ref": "#/components/schemas/domain.UserBalance"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "settlements": {
                        "items": {
                            "$ref": "#/components/schemas/domain.Settlement"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
            "domain.Expense": {
                "properties": {
                    "amount": {
                        "type": "number"
                    },
                    "category": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "createdBy": {
                        "type": "string"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "deletedAt": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "groupId": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "isPayment": {
                        "type": "boolean"
                    },
                    "paidBy": {
                        "type": "string"
                    },
                    "spentAt": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "domain.InputSplit": {
                "properties": {
                    "amount": {
                        "description": "Required if splitType is EXACT",
                        "type": "number"
                    },
                    "percentage": {
                        "description": "Required if splitType is PERCENTAGE",
                        "type": "number"
                    },
                    "userId": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "domain.Settlement": {
                "properties": {
                    "amount": {
                        "type": "number"
                    },
                    "fromUserId": {
                        "type": "string"
                    },
                    "fromUserName": {
                        "type": "string"
                    },
                    "toUserId": {
                        "type": "string"
                    },
                    "toUserName": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "domain.Split": {
                "properties": {
                    "amount": {
                        "type": "number"
                    },
                    "email": {
                        "type": "string"
                    },
                    "expenseId": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "splitType": {
                        "$ref": "#/components/schemas/domain.SplitType"
                    },
                    "splitValue": {
                        "type": "number"
                    },
                    "userId": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "domain.SplitType": {
                "enum": [
                    "EQUAL",
                    "EXACT",
                    "PERCENTAGE"
                ],
                "type": "string",
                "x-enum-varnames": [
                    "SplitTypeEqual",
                    "SplitTypeExact",
                    "SplitTypePercentage"
                ]
            },
...
            "domain.UserBalance": {
                "properties": {
                    "netBalance": {
                        "type": "number"
                    },
                    "userId": {
                        "type": "string"
                    },
                    "userName": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "http.CreateExpenseRequest": {
                "properties": {
                    "amount": {
                        "type": "number"
                    },
                    "category": {
                        "type": "string"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "groupId": {
                        "type": "string"
                    },
                    "paidBy": {
                        "type": "string"
                    },
                    "splitType": {
                        "$ref": "#/components/schemas/domain.SplitType"
                    },
                    "splits": {
                        "items": {
                            "$ref": "#/components/schemas/domain.InputSplit"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
            "http.CreateExpenseResponse": {
                "properties": {
                    "expense": {
                        "$ref": "#/components/schemas/domain.Expense"
                    },
                    "splits": {
                        "items": {
                            "$ref": "#/components/schemas/domain.Split"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
            "http.GetExpenseDetailsResponse": {
                "properties": {
                    "expense": {
                        "$ref": "#/components/schemas/domain.Expense"
                    },
                    "splits": {
                        "items": {
                            "$ref": "#/components/schemas/domain.Split"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
            "http.ListExpensesResponse": {
                "properties": {
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/domain.Expense"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/pagination.Meta"
                    }
                },
                "type": "object"
            },
...
            "http.SettleExpenseRequest": {
                "properties": {
                    "amount": {
                        "type": "number"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "groupId": {
                        "type": "string"
                    },
                    "paidBy": {
                        "type": "string"
                    },
                    "receivedBy": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "http.SettleExpenseResponse": {
                "properties": {
                    "expense": {
                        "$ref": "#/components/schemas/domain.Expense"
                    },
                    "split": {
                        "$ref": "#/components/schemas/domain.Split"
                    }
                },
                "type": "object"
            },
...
            "pagination.Meta": {
                "properties": {
                    "hasMore": {
                        "type": "boolean"
                    },
                    "nextCursor": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "response.ErrorResponse": {
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "Backend API service for Splittr bill splitting.",
        "title": "Splittr API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/balances": {
            "get": {
                "description": "Calculate net balances and recommended settlement transactions.",
                "parameters": [
                    {
                        "description": "Filter by Group ID. If omitted, returns global balances.",
                        "in": "query",
                        "name": "groupId",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Simplify debts algorithm (true/false)",
                        "in": "query",
                        "name": "simplified",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/domain.BalanceResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Get user balances",
                "tags": [
                    "expenses"
                ]
            }
        },
        "/expenses": {
            "get": {
                "description": "Retrieve a cursor-paginated list of expenses filtered by group, personal=true, or friendId.",
                "parameters": [
                    {
                        "description": "Filter by Group ID",
                        "in": "query",
                        "name": "groupId",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter for personal only (true/false)",
                        "in": "query",
                        "name": "personal",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "description": "Filter by Friend ID",
                        "in": "query",
                        "name": "friendId",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Items per page (max 100, default 20)",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Opaque cursor token from a previous response",
                        "in": "query",
                        "name": "cursor",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ListExpensesResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List expenses",
                "tags": [
                    "expenses"
                ]
            },
            "post": {
                "description": "Create a new expense with equal/exact/percentage splits.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/http.CreateExpenseRequest",
                                        "summary": "request",
                                        "description": "Expense details and splits structure"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Expense details and splits structure",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.CreateExpenseResponse"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Create expense",
                "tags": [
                    "expenses"
                ]
            }
        },
        "/expenses/settle": {
            "post": {
                "description": "Create a settlement payment between two users.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/http.SettleExpenseRequest",
                                        "summary": "request",
                                        "description": "Settlement details"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Settlement details",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.SettleExpenseResponse"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Settle balance",
                "tags": [
                    "expenses"
                ]
            }
        },
        "/expenses/{id}": {
            "delete": {
                "description": "Soft-delete an expense by ID.",
                "parameters": [
                    {
                        "description": "Expense ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Delete expense",
                "tags": [
                    "expenses"
                ]
            },
            "get": {
                "description": "Get a specific expense's details including all splits.",
                "parameters": [
                    {
                        "description": "Expense ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.GetExpenseDetailsResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Get expense details",
                "tags": [
                    "expenses"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/v1"
        }
    ]
}